move PagesData type
