+++ packages/frontend/src/pages/bridges-risk/getProps.ts
@@ -1,10 +1,9 @@
 import { Config } from '../../build/config'
-import { PagesData } from '../../build/types'
 import { getFooterProps, getNavbarProps } from '../../components'
 import { getDestination } from '../../utils/getDestination'
 import { getIncludedProjects } from '../../utils/getIncludedProjects'
 import { orderByTvl } from '../../utils/orderByTvl'
-import { Wrapped } from '../Page'
+import { PagesData, Wrapped } from '../Page'
 import { BridgesRiskPageProps } from './BridgesRiskPage'
 import { getPageMetadata } from './getPageMetadata'
 import { BridgesRiskViewEntry } from './types'
