+++ packages/frontend/src/pages/scaling-projects/props/getProps.ts
@@ -1,11 +1,10 @@
 import { Layer2 } from '@l2beat/config'
 
 import { Config } from '../../../build/config'
-import { PagesData } from '../../../build/types'
 import { getFooterProps, getNavbarProps } from '../../../components'
 import { getChart } from '../../../utils/project/getChart'
 import { getHeader } from '../../../utils/project/getHeader'
-import { Wrapped } from '../../Page'
+import { PagesData, Wrapped } from '../../Page'
 import { ProjectPageProps } from '../view/ProjectPage'
 import { getPageMetadata } from './getPageMetadata'
 import { getProjectDetails } from './getProjectDetails'
