+++ packages/frontend/src/pages/scaling-risk/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 
 import { Config } from '../../build/config'
-import { PagesData } from '../../build/types'
 import { PageWrapper } from '../../components'
+import { PagesData } from '../Page'
 import { getProps } from './props'
 import { ScalingRiskPage } from './view/ScalingRiskPage'
