+++ packages/frontend/src/pages/scaling-risk/props/getProps.ts
@@ -1,9 +1,8 @@
 import { Config } from '../../../build/config'
-import { PagesData } from '../../../build/types'
 import { getFooterProps, getNavbarProps } from '../../../components'
 import { getIncludedProjects } from '../../../utils/getIncludedProjects'
 import { orderByTvl } from '../../../utils/orderByTvl'
-import { Wrapped } from '../../Page'
+import { PagesData, Wrapped } from '../../Page'
 import { ScalingRiskPageProps } from '../view/ScalingRiskPage'
 import { getPageMetadata } from './getPageMetadata'
 import { getRiskView } from './getRiskView'
