+++ packages/frontend/src/pages/scaling-tvl/props/getProps.ts
@@ -1,11 +1,10 @@
 import { Config } from '../../../build/config'
-import { PagesData } from '../../../build/types'
 import { getFooterProps, getNavbarProps } from '../../../components'
 import { getIncludedProjects } from '../../../utils/getIncludedProjects'
 import { orderByTvl } from '../../../utils/orderByTvl'
 import { getTvlWithChange } from '../../../utils/tvl/getTvlWitchChange'
 import { formatUSD } from '../../../utils/utils'
-import { Wrapped } from '../../Page'
+import { PagesData, Wrapped } from '../../Page'
 import { TvlPageProps } from '../view/ScalingTvlPage'
 import { getPageMetadata } from './getPageMetadata'
 import { getScalingTvlView } from './getScalingTvlView'
